Modify import statements to reflect refactored package name
