+++ src/IntegrationTests/java/uk/gov/hmcts/sscs/sya/SyaEndpointsIt.java
@@ -55,9 +55,9 @@ import uk.gov.hmcts.reform.sscs.ccd.client.CcdClient;
 import uk.gov.hmcts.reform.sscs.idam.Authorize;
 import uk.gov.hmcts.reform.sscs.idam.IdamApiClient;
 import uk.gov.hmcts.reform.sscs.idam.UserDetails;
-import uk.gov.hmcts.sscs.controller.SyaController;
-import uk.gov.hmcts.sscs.domain.wrapper.SyaCaseWrapper;
-import uk.gov.hmcts.sscs.model.pdf.PdfWrapper;
+import uk.gov.hmcts.reform.sscs.controller.SyaController;
+import uk.gov.hmcts.reform.sscs.domain.wrapper.SyaCaseWrapper;
+import uk.gov.hmcts.reform.sscs.model.pdf.PdfWrapper;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
