+++ src/e2e/java/uk/gov/hmcts/sscs/functional/evidence/EvidenceDocumentUploadTest.java
@@ -15,8 +15,8 @@ import java.util.ArrayList;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.junit.Test;
-import uk.gov.hmcts.sscs.domain.wrapper.SyaCaseWrapper;
-import uk.gov.hmcts.sscs.domain.wrapper.SyaEvidence;
+import uk.gov.hmcts.reform.sscs.domain.wrapper.SyaCaseWrapper;
+import uk.gov.hmcts.reform.sscs.domain.wrapper.SyaEvidence;
 
 
 public class EvidenceDocumentUploadTest {
