+++ src/main/java/uk/gov/hmcts/sscs/controller/EvidenceManagementController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 import uk.gov.hmcts.reform.document.domain.UploadResponse;
 import uk.gov.hmcts.reform.sscs.service.EvidenceManagementService;
-import uk.gov.hmcts.sscs.exception.EvidenceDocumentsMissingException;
+import uk.gov.hmcts.reform.sscs.exception.EvidenceDocumentsMissingException;
 
 @RestController
 public class EvidenceManagementController {
