+++ src/test/java/uk/gov/hmcts/sscs/controller/EvidenceManagementControllerTest.java
@@ -13,7 +13,7 @@ import org.mockito.Mock;
 import org.springframework.web.multipart.MultipartFile;
 import uk.gov.hmcts.reform.document.domain.UploadResponse;
 import uk.gov.hmcts.reform.sscs.service.EvidenceManagementService;
-import uk.gov.hmcts.sscs.exception.EvidenceDocumentsMissingException;
+import uk.gov.hmcts.reform.sscs.exception.EvidenceDocumentsMissingException;
 
 
 public class EvidenceManagementControllerTest {
