+++ src/test/java/uk/gov/hmcts/sscs/service/TribunalsServiceTest.java
@@ -22,12 +22,12 @@ import uk.gov.hmcts.reform.sscs.ccd.exception.CcdException;
 import uk.gov.hmcts.reform.sscs.ccd.service.CcdService;
 import uk.gov.hmcts.reform.sscs.idam.IdamService;
 import uk.gov.hmcts.reform.sscs.idam.IdamTokens;
-import uk.gov.hmcts.sscs.builder.TrackYourAppealJsonBuilder;
-import uk.gov.hmcts.sscs.exception.AppealNotFoundException;
-import uk.gov.hmcts.sscs.model.tya.SubscriptionRequest;
-import uk.gov.hmcts.sscs.model.tya.SurnameResponse;
-import uk.gov.hmcts.sscs.service.exceptions.InvalidSurnameException;
-import uk.gov.hmcts.sscs.service.referencedata.RegionalProcessingCenterService;
+import uk.gov.hmcts.reform.sscs.builder.TrackYourAppealJsonBuilder;
+import uk.gov.hmcts.reform.sscs.exception.AppealNotFoundException;
+import uk.gov.hmcts.reform.sscs.model.tya.SubscriptionRequest;
+import uk.gov.hmcts.reform.sscs.model.tya.SurnameResponse;
+import uk.gov.hmcts.reform.sscs.service.exceptions.InvalidSurnameException;
+import uk.gov.hmcts.reform.sscs.service.referencedata.RegionalProcessingCenterService;
 
 @RunWith(MockitoJUnitRunner.class)
 public class TribunalsServiceTest {
