+++ src/test/java/uk/gov/hmcts/sscs/transform/deserialize/SubmitYourAppealToCcdCaseDataDeserializerTest.java
@@ -10,7 +10,7 @@ import org.junit.Before;
 import org.junit.Test;
 import uk.gov.hmcts.reform.sscs.ccd.domain.RegionalProcessingCenter;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
-import uk.gov.hmcts.sscs.domain.wrapper.SyaCaseWrapper;
+import uk.gov.hmcts.reform.sscs.domain.wrapper.SyaCaseWrapper;
 
 public class SubmitYourAppealToCcdCaseDataDeserializerTest {
